+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project are documented in this file.
 ### Added
 
 - PWA export support
+- Custom leaderboard system for PWA version
 
 ### Fixed